// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  PRESIDENT: 'PRESIDENT',
  EXECUTIVE: 'EXECUTIVE',
  DIRECTOR: 'DIRECTOR',
  MANAGER: 'MANAGER',
  LEADER: 'LEADER',
  STAFF: 'STAFF'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const EvaluationStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED_TO_LEADER: 'SUBMITTED_TO_LEADER',
  SUBMITTED_TO_MANAGER: 'SUBMITTED_TO_MANAGER',
  SUBMITTED_TO_DIRECTOR: 'SUBMITTED_TO_DIRECTOR',
  SUBMITTED_TO_EXECUTIVE: 'SUBMITTED_TO_EXECUTIVE',
  SUBMITTED_TO_PRESIDENT: 'SUBMITTED_TO_PRESIDENT',
  COMPLETED: 'COMPLETED'
} as const

export type EvaluationStatus = (typeof EvaluationStatus)[keyof typeof EvaluationStatus]